Reorganized backend project, removed DI

+++ backend/src/main/scala/com/pp/students_organizer_backend/domain/AssignmentTypeEntity.scala
@@ -1,5 +1,5 @@
 package com.pp.students_organizer_backend.domain
 
-case class AssignmentTypeEntity(id: Long,
+case class AssignmentTypeEntity(id: Int,
                                 name: String,
                                )
